Check if servecm (Chart Manager) is already installed.
Issue-ID: AIMLFW-43

+++ bin/install_common_templates_to_helm.sh
@@ -17,12 +17,16 @@
 # ==================================================================================
 
 #!/bin/bash
+#set -e
 
-echo "Installing servecm (Chart Manager) and common templates to helm3"
+if [ -e $(helm plugin list| grep servecm | awk '{print $1}') ];
+then
+	echo "Installing servecm (Chart Manager) and common templates to helm3"
 
-helm plugin install https://github.com/jdolitsky/helm-servecm
-eval $(helm env |grep HELM_REPOSITORY_CACHE) 
-echo ${HELM_REPOSITORY_CACHE}
+	helm plugin install https://github.com/jdolitsky/helm-servecm
+	eval $(helm env | grep HELM_REPOSITORY_CACHE)
+	echo ${HELM_REPOSITORY_CACHE}
+fi
 
 #The hardcoded location does not work anymore, Reference: change similar to RIC-966 updating chartmuseum and docker refs
 curl -LO https://get.helm.sh/chartmuseum-v0.15.0-linux-386.tar.gz
